Group Fixes
+ Started Rewriting Group
+ Cleaned Table groups_data
+ Created table groups_forums_data
+ Improved Group Forum System
+ Fixed Buy Group
+ Fixed DC when enter room
+ Improved some things

=> Bugs
+ Create Thread and Open Thread gives DC (Wrong Packet)
+ Diffielman Weirdly isn't working... (Only for non-RC4 patched
Habbo.swf's)

+++ Yupi/Game/Users/Habbo.cs
@@ -169,7 +169,7 @@ namespace Yupi.Game.Users
         /// <summary>
         ///     The favourite group
         /// </summary>
-        internal int FavouriteGroup;
+        internal uint FavouriteGroup;
 
         /// <summary>
         ///     The flood time
@@ -548,7 +548,7 @@ namespace Yupi.Game.Users
             _currentQuestProgress = currentQuestProgress;
             IsHopping = false;
 
-            FavouriteGroup = Yupi.GetGame().GetGroupManager().GetGroup(favId) != null ? favId : 0;
+            FavouriteGroup = Yupi.GetGame().GetGroupManager().GetGroup((uint) favId) != null ? (uint)favId : 0;
 
             UserGroups = groups;
